Namespace function to_toml()
