+++ lib/puppet/functions/to_toml.rb
@@ -1,22 +1,14 @@
 # frozen_string_literal: true
 
-require_relative '../../puppet_x/stdlib/toml_dumper'
+# THIS FILE WAS GENERATED BY `rake regenerate_unamespaced_shims`
 
-# @summary Convert a data structure and output to TOML.
+# @summary DEPRECATED.  Use the namespaced function [`stdlib::to_toml`](#stdlibto_toml) instead.
 Puppet::Functions.create_function(:to_toml) do
-  # @param data Data structure which needs to be converted into TOML
-  # @return [String] Converted data as TOML string
-  # @example How to output TOML to a file
-  #     file { '/tmp/config.toml':
-  #       ensure  => file,
-  #       content => to_toml($myhash),
-  #     }
-  dispatch :to_toml do
-    required_param 'Hash', :data
-    return_type 'String'
+  dispatch :deprecation_gen do
+    repeated_param 'Any', :args
   end
-
-  def to_toml(data)
-    PuppetX::Stdlib::TomlDumper.new(data).toml_str
+  def deprecation_gen(*args)
+    call_function('deprecation', 'to_toml', 'This function is deprecated, please use stdlib::to_toml instead.')
+    call_function('stdlib::to_toml', *args)
   end
 end
