+++ spec/functions/to_toml_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe 'to_toml' do
+describe 'stdlib::to_toml' do
   context 'fails on invalid params' do
     it { is_expected.not_to be_nil }
